Add files via upload
The Employee Payroll Management System is a Java-based application designed to manage employee data and automate the payroll process. This system simplifies the management of employee information, calculates salaries based on work hours, generates payroll reports. It is user-friendly, efficient, and provides accurate payroll processing for small to medium-sized organizations